Move `parse_osu_config` to `error` route
