+++ app/routes/web/error.py
@@ -1,9 +1,11 @@
 
-from app.common.database.repositories import logs, users
+from __future__ import annotations
+
+from app.common.database import users
 from app.common import officer
 
 from sqlalchemy.orm import Session
-from typing import Optional
+from typing import Dict
 from fastapi import (
     HTTPException,
     APIRouter,
@@ -15,31 +17,39 @@ from fastapi import (
 router = APIRouter()
 
 import bcrypt
-import utils
 import json
 import app
+
+def parse_osu_config(config: str) -> Dict[str, str]:
+    return {
+        k.strip():v.strip()
+        for (k, v) in [
+            line.split('=', 1) for line in config.splitlines()
+            if '=' in line and not line.startswith('#')
+        ]
+    }
 
 @router.post('/osu-error.php')
 def osu_error(
     session: Session = Depends(app.session.database.yield_session),
     username: str = Form(..., alias='u'),
-    user_id: Optional[int] = Form(None, alias='i'),
+    user_id: int | None = Form(None, alias='i'),
     language: str = Form(..., alias='culture'),
     mode: str = Form(..., alias='gamemode'),
     time: int = Form(..., alias='gametime'),
-    beatmap_id: Optional[int] = Form(None, alias='b'),
-    beatmap_md5: Optional[str] = Form(None, alias='bc'),
+    beatmap_id: int | None = Form(None, alias='b'),
+    beatmap_md5: str | None = Form(None, alias='bc'),
     audiotime: int = Form(...),
     exception: str = Form(...),
     stacktrace: str = Form(...),
-    feedback: Optional[str] = Form(None),
-    iltrace: Optional[str] = Form(None),
-    exehash: Optional[str] = Form(None),
+    feedback: str | None = Form(None),
+    iltrace: str | None = Form(None),
+    exehash: str | None = Form(None),
     version: str = Form(...),
     config: str = Form(...)
 ):
     # Parse config to get password
-    config = utils.parse_osu_config(config)
+    config = parse_osu_config(config)
 
     if not (user := users.fetch_by_id(user_id, session=session)):
         raise HTTPException(401)
